separate tests from normal source files
